relax types for Bind and Render funcs
This relaxes the Bind and Render to accept any type as a payload.  If
those types implement the Binder or Renderer interfaces, then the
relevant method on the type is called.  Sometimes the default Bind or
Render behavior is sufficient, and I don't need any additional custom
handling. Return early if payload value is nil.

Signed-off-by: Will Norris <will@tailscale.com>